Initial attempt at documentation (using Sphinx as prepared by PyScaffold) appears to be somewhat successful.
Added **.. include:: README.rst** to index.rst to avoid
duplication of the documentation. This made lead to problems
on GitHub later, but a problem delayed is a problem denied. :)

TODO: Replace "skeleton.py" with a function that stands up
a new project. This should allow the user to specify an
initial set of processes that will be used to populate a file
hierarachy and a tracking.xlsx file. (This may be replaced by
a SQL database at some point.)

TODO: Add forward and reverse inheritance operations, such as
property reductions and property-to-condition mappings when
passing down properties from parent to child.

TODO: Writer for Citrine JSON.