"projects" directory, which lies outside the src directory, and therefore outside the karon package, is meant to prototype the use of this library in quasi-realistic development scenarios. A separate set of tests are present in this directory. These tests and all tests for the karon package pass.
TODO: replace skeleton.py with a CLI that allows the user to
stand up a new project with an associated directory hierarchy and
data metadata tables. These will start with simple Excel, but
can, and should, quickly extend to SQL.

TODO: Connect to Citrine PIF.

TODO: Develop (in a feature branch) structures to facilitate the
construction of Samples. The Example object in example.py works,
and is not too clumsy. IDEA: Use Example to build a base Walker/
Agent. Inherit from this to create a `PropagateAgent` and a
`AggregateAgent`.

TODO: Reader. The Example object in example.py conflates a reader,
an aggregate, and a propagate agent. The last two are addressed
above. Still need a reader.