Refactoring is complete, and karon is now functional, at least in its functionality applied to `projects/qualitymade.py`. While relatively few additions/deletions were necessary, more are likely as common templates are identified in the workflow.
The anticipated use case is for the contents of each node to
contain attributes, stored as a dictionary. Two function
generators--`get` and `put`--can be used to read and write
to specific keys in *contents*, assuming *contents* exposes
a __getitem__ and __setitem__ method.

Both project and package tests pass.

TODO: replace skeleton.py with a CLI that allows the user to
stand up a new project with an associated directory hierarchy and
data metadata tables. These will start with simple Excel, but
can, and should, quickly extend to SQL.

TODO: Connect to Citrine PIF.

TODO: Develop (in a feature branch) structures to facilitate the
construction of Samples. The Example object in example.py works,
and is not too clumsy. IDEA: Use Example to build a base Walker/
Agent. Inherit from this to create a `PropagateAgent` and a
`AggregateAgent`.

TODO: Reader. The Example object in example.py conflates a reader,
an aggregate, and a propagate agent. The last two are addressed
above. Still need a reader.