Refined the series of functions that are applied to the tree sparked by a specific case study.
JP (Lincoln Electric Additive Technologies) records nominal
build parameter settings on a per-layer basis. Therefore,
a single part (for example, a wall) will have multiple
settings. These are collected into a vector. If a value
is, itself, vector-valued, then aggregation of these will
create a list-of-lists.

1. These lists are flattened.
2. Statistics are calculated from each vector-valued
   property: min, max, mean, std.

The ability to calculate new fields based on existing
data is also included. The first case study for this
capability is the "remelt ratio" from Jack Beuth. There
are a few critical design considerations:

1. This requires an a priori knowledge of the data column
   headers.
  1. Currently, this is captured by a series of calls to
     check for specific columns.
  2. TODO: This should be defined in a configuration file,
     to allow the user to set the headers.
2. One, or more, necessary fields may be missing. What
   should be done when that data is missing: abort? skip?
   default?